backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.5.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
google-generativeai==0.7.2
xrpl-py==2.6.0
requests==2.31.0
aiohttp==3.9.1
reportlab==4.0.7
Pillow==10.1.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
pyngrok==7.2.0
python-multipart==0.0.9
httpx==0.27.0
cryptography==42.0.0
